refactor: split EvalFull
