+++ primer/primer.cabal
@@ -73,8 +73,10 @@ library
     Primer.Eval.EvalError
     Primer.Eval.Inline
     Primer.Eval.Let
+    Primer.Eval.NormalOrder
     Primer.Eval.Prim
     Primer.Eval.Push
+    Primer.Eval.Redex
     Primer.Eval.Utils
     Primer.Primitives.PrimDef
     Primer.Typecheck.Cxt
